Multitracker.py run refactor
- Attempt to refactor run command

This breaks up the run function into code grouped together, and often into triggers that are only ran if a certain condition is met. Each of these are broken up. The trackers are then processed, and statistical filters are ran.
The common values between all of these functions are now class variables.

Nothing is replaced and none of this is tested.

Todo: test and integrate the new code, and document it further.